refactor: remove AttioIntegration implementation and add tests
- Deleted the existing AttioIntegration implementation file.
- Created a new test file for AttioIntegration to ensure proper functionality and integration with BaseCRMIntegration.
- Added tests for static configuration, required methods, backward compatibility, and helper methods.
- Updated package-lock.json to include the latest version of @friggframework/api-module-attio.
