+++ backend/.env.example
@@ -9,14 +9,17 @@ MONGO_URI=mongodb://root:rootpassword@localhost:27017?retryWrites=true&w=majorit
 #BASE_URL=CUSTOM BASE URL
 
 
-ASANA_CLIENT_ID=
-ASANA_CLIENT_SECRET=
-ASANA_SCOPE=default openid
-SKIP_ASANA_SIGNATURE_VALIDATION=false
+# Quo CRM API
+QUO_API_KEY=
+SCALE_TEST_API_KEY=
 
-FRONTIFY_CLIENT_ID=
-FRONTIFY_SCOPE="basic:read basic:write finder:read account:read webhook:read webhook:write"
-HEALTH_API_KEY=
+#Axis Care
+AXISCARE_API_KEY=
+AXISCARE_BASE_URL=
 
+#Attio
+ATTIO_CLIENT_ID=
+ATTIO_CLIENT_SECRET=
+ATTIO_SCOPE=user_management:read record_permission:read-write
