+++ backend/src/integrations/AttioIntegration.js
@@ -1,6 +1,12 @@
-const { IntegrationBase } = require('@friggframework/core');
+const { BaseCRMIntegration } = require('../base/BaseCRMIntegration');
+const attio = require('@friggframework/api-module-attio');
+const quo = require('../api-modules/quo');
 
-class AttioIntegration extends IntegrationBase {
+/**
+ * Attio-specific implementation for syncing people/companies with Quo.
+ * Attio uses a modern record-based API with flexible data structures.
+ */
+class AttioIntegration extends BaseCRMIntegration {
     static Definition = {
         name: 'attio',
         version: '1.0.0',
@@ -15,12 +21,8 @@ class AttioIntegration extends IntegrationBase {
             icon: '',
         },
         modules: {
-            attio: {
-                definition: require('/Users/danielklotz/projects/lefthook/api-module-library/packages/v1-ready/attio').Definition,
-            },
-            quo: {
-                definition: require('../api-modules/quo').Definition,
-            },
+            attio: { definition: attio.Definition },
+            quo: { definition: quo.Definition },
         },
         routes: [
             {
@@ -46,9 +48,37 @@ class AttioIntegration extends IntegrationBase {
         ],
     };
 
-    constructor() {
-        super();
+    /**
+     * CRM Configuration - Required by BaseCRMIntegration
+     */
+    static CRMConfig = {
+        personObjectTypes: [
+            { crmObjectName: 'people', quoContactType: 'contact' },
+        ],
+        syncConfig: {
+            reverseChronological: true,
+            initialBatchSize: 50,
+            ongoingBatchSize: 25,
+            supportsWebhooks: true, // Attio has webhook support
+            pollIntervalMinutes: 30,
+        },
+        queueConfig: {
+            maxWorkers: 15,
+            provisioned: 5,
+            maxConcurrency: 50,
+            batchSize: 1,
+            timeout: 600,
+        },
+    };
+
+    constructor(params) {
+        super(params);
+
+        // Add existing events (backward compatibility)
         this.events = {
+            ...this.events, // BaseCRMIntegration events
+
+            // Existing Attio-specific events
             LIST_ATTIO_WORKSPACES: {
                 handler: this.listWorkspaces,
             },
@@ -68,20 +98,6 @@ class AttioIntegration extends IntegrationBase {
                 description: 'Retrieve custom object types from Attio',
                 userActionType: 'DATA',
             },
-            SYNC_COMPANIES_TO_QUO: {
-                type: 'USER_ACTION',
-                handler: this.syncCompaniesToQuo,
-                title: 'Sync Companies to Quo',
-                description: 'Synchronize Attio companies with Quo CRM',
-                userActionType: 'DATA',
-            },
-            SYNC_PEOPLE_TO_QUO: {
-                type: 'USER_ACTION',
-                handler: this.syncPeopleToQuo,
-                title: 'Sync People to Quo',
-                description: 'Synchronize Attio people with Quo CRM',
-                userActionType: 'DATA',
-            },
             CREATE_ATTIO_RECORD: {
                 type: 'USER_ACTION',
                 handler: this.createRecord,
@@ -99,23 +115,297 @@ class AttioIntegration extends IntegrationBase {
         };
     }
 
+    // ============================================================================
+    // REQUIRED METHODS - BaseCRMIntegration Abstract Methods
+    // ============================================================================
+
+    /**
+     * Fetch a page of persons from Attio
+     * @param {Object} params
+     * @param {string} params.objectType - CRM object type (people)
+     * @param {number} params.page - Page number (0-indexed)
+     * @param {number} params.limit - Records per page
+     * @param {Date} [params.modifiedSince] - Filter by modification date
+     * @param {boolean} [params.sortDesc=true] - Sort descending
+     * @returns {Promise<{data: Array, total: number, hasMore: boolean}>}
+     */
+    async fetchPersonPage({
+        objectType,
+        page,
+        limit,
+        modifiedSince,
+        sortDesc = true,
+    }) {
+        try {
+            const params = {
+                limit,
+                offset: page * limit,
+                sort: {
+                    attribute: 'updated_at',
+                    direction: sortDesc ? 'desc' : 'asc',
+                },
+            };
+
+            // Add modification filter if provided
+            if (modifiedSince) {
+                params.filter = {
+                    attribute: 'updated_at',
+                    gte: modifiedSince.toISOString(),
+                };
+            }
+
+            // Attio uses object slugs (e.g., 'people', 'companies')
+            const response = await this.attio.api.objects.listRecords(
+                objectType,
+                params,
+            );
+
+            return {
+                data: response.data || [],
+                total: response.total || null,
+                hasMore: response.has_more || false,
+            };
+        } catch (error) {
+            console.error(`Error fetching ${objectType} page ${page}:`, error);
+            throw error;
+        }
+    }
+
+    /**
+     * Transform Attio person object to Quo contact format
+     * @param {Object} person - Attio person record
+     * @returns {Promise<Object>} Quo contact format
+     */
+    async transformPersonToQuo(person) {
+        // Attio uses a flexible attribute-based structure
+        const attributes = person.values || {};
+
+        // Extract name from attributes
+        const nameAttr = attributes.name?.[0];
+        const firstName = nameAttr?.first_name || '';
+        const lastName = nameAttr?.last_name || '';
+
+        // Extract email addresses
+        const emails = [];
+        const emailAttrs = attributes.email_addresses || [];
+        for (const emailAttr of emailAttrs) {
+            if (emailAttr.email_address) {
+                emails.push({
+                    name: emailAttr.attribute_type || 'work',
+                    value: emailAttr.email_address,
+                    primary: emailAttr.is_primary || false,
+                });
+            }
+        }
+
+        // Extract phone numbers
+        const phoneNumbers = [];
+        const phoneAttrs = attributes.phone_numbers || [];
+        for (const phoneAttr of phoneAttrs) {
+            if (phoneAttr.phone_number) {
+                phoneNumbers.push({
+                    name: phoneAttr.attribute_type || 'work',
+                    value: phoneAttr.phone_number,
+                    primary: phoneAttr.is_primary || false,
+                });
+            }
+        }
+
+        // Extract company (from primary company relationship)
+        let company = null;
+        const companyLinks = attributes.primary_company || [];
+        if (companyLinks.length > 0 && companyLinks[0].target_record_id) {
+            try {
+                const companyRecord = await this.attio.api.objects.getRecord(
+                    'companies',
+                    companyLinks[0].target_record_id,
+                );
+                const companyName = companyRecord.values?.name?.[0]?.value;
+                company = companyName || null;
+            } catch (error) {
+                console.warn(
+                    `Failed to fetch company for person ${person.id.record_id}:`,
+                    error.message,
+                );
+            }
+        }
+
+        return {
+            externalId: person.id.record_id,
+            source: 'attio',
+            defaultFields: {
+                firstName,
+                lastName,
+                company,
+                phoneNumbers,
+                emails,
+            },
+            customFields: {
+                crmId: person.id.record_id,
+                crmType: 'attio',
+                objectId: person.id.object_id,
+                createdAt: person.created_at,
+                updatedAt: person.updated_at,
+                // Store all Attio attributes for reference
+                attioAttributes: attributes,
+            },
+        };
+    }
+
+    /**
+     * Log SMS message to Attio as a note/interaction
+     * @param {Object} activity - SMS activity
+     * @returns {Promise<void>}
+     */
+    async logSMSToActivity(activity) {
+        try {
+            // Find the person by external ID
+            const person = await this.attio.api.objects.getRecord(
+                'people',
+                activity.contactExternalId,
+            );
+            if (!person) {
+                console.warn(
+                    `Person not found for SMS logging: ${activity.contactExternalId}`,
+                );
+                return;
+            }
+
+            // Create note entry in Attio
+            const noteData = {
+                parent_object: 'people',
+                parent_record_id: activity.contactExternalId,
+                title: `SMS: ${activity.direction}`,
+                content: activity.content,
+                created_at: activity.timestamp,
+            };
+
+            await this.attio.api.notes.create(noteData);
+        } catch (error) {
+            console.error('Failed to log SMS activity to Attio:', error);
+            throw error;
+        }
+    }
+
+    /**
+     * Log phone call to Attio as a note/interaction
+     * @param {Object} activity - Call activity
+     * @returns {Promise<void>}
+     */
+    async logCallToActivity(activity) {
+        try {
+            // Find the person by external ID
+            const person = await this.attio.api.objects.getRecord(
+                'people',
+                activity.contactExternalId,
+            );
+            if (!person) {
+                console.warn(
+                    `Person not found for call logging: ${activity.contactExternalId}`,
+                );
+                return;
+            }
+
+            // Create note entry in Attio
+            const noteData = {
+                parent_object: 'people',
+                parent_record_id: activity.contactExternalId,
+                title: `Call: ${activity.direction} (${activity.duration}s)`,
+                content: activity.summary || 'Phone call',
+                created_at: activity.timestamp,
+            };
+
+            await this.attio.api.notes.create(noteData);
+        } catch (error) {
+            console.error('Failed to log call activity to Attio:', error);
+            throw error;
+        }
+    }
+
+    /**
+     * Setup webhooks with Attio
+     * @returns {Promise<void>}
+     */
+    async setupWebhooks() {
+        try {
+            const webhookUrl = `${process.env.BASE_URL}/integrations/${this.id}/webhook`;
+
+            // Create webhook for record.created events
+            await this.attio.api.webhooks.create({
+                url: webhookUrl,
+                subscribed_events: [
+                    'record.created',
+                    'record.updated',
+                    'record.deleted',
+                ],
+                object_types: ['people'],
+            });
+
+            console.log(`Attio webhooks created for integration ${this.id}`);
+        } catch (error) {
+            console.error('Failed to setup Attio webhooks:', error);
+            // Don't throw - fallback to polling
+        }
+    }
+
+    // ============================================================================
+    // OPTIONAL HELPER METHODS
+    // ============================================================================
+
+    /**
+     * Fetch a single person by ID
+     * @param {string} id - Person ID
+     * @returns {Promise<Object>}
+     */
+    async fetchPersonById(id) {
+        return await this.attio.api.objects.getRecord('people', id);
+    }
+
+    /**
+     * Fetch multiple persons by IDs (for webhook batch processing)
+     * @param {string[]} ids - Array of person IDs
+     * @returns {Promise<Object[]>}
+     */
+    async fetchPersonsByIds(ids) {
+        const persons = [];
+        for (const id of ids) {
+            try {
+                const person = await this.fetchPersonById(id);
+                persons.push(person);
+            } catch (error) {
+                console.error(`Failed to fetch person ${id}:`, error.message);
+            }
+        }
+        return persons;
+    }
+
+    // ============================================================================
+    // EXISTING METHODS - Backward Compatibility
+    // ============================================================================
+
     async listWorkspaces({ req, res }) {
         try {
-            const workspaces = await this.attio.api.listWorkspaces();
+            const workspaces = await this.attio.api.workspaces.list();
             res.json(workspaces);
         } catch (error) {
             console.error('Failed to list Attio workspaces:', error);
-            res.status(500).json({ error: 'Failed to list workspaces', details: error.message });
+            res.status(500).json({
+                error: 'Failed to list workspaces',
+                details: error.message,
+            });
         }
     }
 
     async listObjects({ req, res }) {
         try {
-            const objects = await this.attio.api.listObjectTypes();
+            const objects = await this.attio.api.objects.list();
             res.json(objects);
         } catch (error) {
             console.error('Failed to list Attio objects:', error);
-            res.status(500).json({ error: 'Failed to list objects', details: error.message });
+            res.status(500).json({
+                error: 'Failed to list objects',
+                details: error.message,
+            });
         }
     }
 
@@ -123,15 +413,20 @@ class AttioIntegration extends IntegrationBase {
         try {
             const params = {
                 limit: req.query.limit ? parseInt(req.query.limit) : 50,
-                cursor: req.query.cursor,
-                filter: req.query.filter ? JSON.parse(req.query.filter) : null,
+                offset: req.query.offset ? parseInt(req.query.offset) : 0,
             };
 
-            const companies = await this.attio.api.listCompanies(params);
+            const companies = await this.attio.api.objects.listRecords(
+                'companies',
+                params,
+            );
             res.json(companies);
         } catch (error) {
             console.error('Failed to list Attio companies:', error);
-            res.status(500).json({ error: 'Failed to list companies', details: error.message });
+            res.status(500).json({
+                error: 'Failed to list companies',
+                details: error.message,
+            });
         }
     }
 
@@ -139,666 +434,86 @@ class AttioIntegration extends IntegrationBase {
         try {
             const params = {
                 limit: req.query.limit ? parseInt(req.query.limit) : 50,
-                cursor: req.query.cursor,
-                filter: req.query.filter ? JSON.parse(req.query.filter) : null,
+                offset: req.query.offset ? parseInt(req.query.offset) : 0,
             };
 
-            const people = await this.attio.api.listPeople(params);
+            const people = await this.attio.api.objects.listRecords(
+                'people',
+                params,
+            );
             res.json(people);
         } catch (error) {
             console.error('Failed to list Attio people:', error);
-            res.status(500).json({ error: 'Failed to list people', details: error.message });
+            res.status(500).json({
+                error: 'Failed to list people',
+                details: error.message,
+            });
         }
     }
 
-    async getCustomObjects(args) {
+    async getCustomObjects({ req, res }) {
         try {
-            // Get all object types
-            const objectTypes = await this.attio.api.listObjectTypes();
+            const objects = await this.attio.api.objects.list();
+            // Filter to only custom objects (not standard ones like 'people', 'companies')
+            const customObjects = objects.data?.filter(
+                (obj) => !['people', 'companies'].includes(obj.api_slug),
+            );
+            res.json({ data: customObjects });
+        } catch (error) {
+            console.error('Failed to get Attio custom objects:', error);
+            res.status(500).json({
+                error: 'Failed to get custom objects',
+                details: error.message,
+            });
+        }
+    }
 
-            // Filter for custom objects (non-standard ones)
-            const customObjects = objectTypes.data?.filter(obj =>
-                !['company', 'person', 'workspace'].includes(obj.object_type)
-            ) || [];
+    async createRecord({ req, res }) {
+        try {
+            const { objectType, values } = req.body;
 
-            // Get sample records for each custom object
-            const objectData = [];
-
-            for (const objectType of customObjects.slice(0, args.maxObjects || 5)) {
-                try {
-                    const records = await this.attio.api.listRecords(objectType.object_type, {
-                        limit: args.sampleLimit || 3,
-                    });
-
-                    objectData.push({
-                        objectType: objectType.object_type,
-                        displayName: objectType.display_name,
-                        description: objectType.description,
-                        properties: objectType.properties,
-                        sampleRecords: records.data?.slice(0, args.sampleLimit || 3) || [],
-                        recordCount: records.meta?.count || 0,
-                    });
-                } catch (error) {
-                    objectData.push({
-                        objectType: objectType.object_type,
-                        displayName: objectType.display_name,
-                        error: error.message,
-                        sampleRecords: [],
-                    });
-                }
+            if (!objectType || !values) {
+                return res.status(400).json({
+                    error: 'objectType and values are required',
+                });
             }
 
-            return {
-                label: 'Attio Custom Objects',
-                data: {
-                    summary: {
-                        totalCustomObjects: customObjects.length,
-                        objectsWithSampleData: objectData.filter(obj => !obj.error).length,
-                        totalRecords: objectData.reduce((sum, obj) => sum + (obj.recordCount || 0), 0),
-                    },
-                    objectData,
-                    timestamp: new Date().toISOString(),
-                }
-            };
+            const result = await this.attio.api.objects.createRecord(
+                objectType,
+                { values },
+            );
+            res.json(result);
         } catch (error) {
-            console.error('Failed to get custom objects:', error);
-            throw new Error(`Failed to get custom objects: ${error.message}`);
+            console.error('Failed to create Attio record:', error);
+            res.status(500).json({
+                error: 'Failed to create record',
+                details: error.message,
+            });
         }
     }
 
-    async syncCompaniesToQuo(args) {
+    async searchRecords({ req, res }) {
         try {
-            // Get companies from Attio
-            const attioCompanies = await this.attio.api.listCompanies({
-                limit: args.limit || 50,
-                cursor: args.cursor,
-                filter: args.filter ? JSON.parse(args.filter) : null,
+            const { query, object_types } = req.body;
+
+            if (!query) {
+                return res.status(400).json({
+                    error: 'Search query is required',
+                });
+            }
+
+            const result = await this.attio.api.search({
+                query,
+                object_types: object_types || ['people', 'companies'],
             });
 
-            const syncResults = [];
-
-            for (const company of attioCompanies.data?.slice(0, args.maxCompanies || 10) || []) {
-                try {
-                    // Transform company data for Quo
-                    const quoCompanyData = await this.transformCompanyForQuo(company);
-
-                    // Create or update in Quo if available
-                    let quoResult = null;
-                    if (this.quo?.api) {
-                        quoResult = await this.createQuoCompany(quoCompanyData);
-                    }
-
-                    syncResults.push({
-                        attioCompany: {
-                            id: company.id,
-                            name: this.getCompanyName(company),
-                            domain: this.getCompanyDomain(company),
-                            status: company.company_status?.values?.[0]?.value,
-                        },
-                        quoCompany: quoResult,
-                        syncStatus: quoResult ? 'success' : 'quo_unavailable',
-                        timestamp: new Date().toISOString(),
-                    });
-                } catch (companyError) {
-                    syncResults.push({
-                        attioCompany: company,
-                        error: companyError.message,
-                        syncStatus: 'error',
-                        timestamp: new Date().toISOString(),
-                    });
-                }
-            }
-
-            return {
-                label: 'Company Sync Results',
-                data: {
-                    totalCompaniesProcessed: syncResults.length,
-                    syncSummary: syncResults.reduce((summary, result) => {
-                        summary[result.syncStatus] = (summary[result.syncStatus] || 0) + 1;
-                        return summary;
-                    }, {}),
-                    syncResults,
-                    timestamp: new Date().toISOString(),
-                }
-            };
+            res.json(result);
         } catch (error) {
-            console.error('Company sync failed:', error);
-            throw new Error(`Company sync failed: ${error.message}`);
-        }
-    }
-
-    async syncPeopleToQuo(args) {
-        try {
-            // Get people from Attio
-            const attioPeople = await this.attio.api.listPeople({
-                limit: args.limit || 50,
-                cursor: args.cursor,
-                filter: args.filter ? JSON.parse(args.filter) : null,
+            console.error('Failed to search Attio records:', error);
+            res.status(500).json({
+                error: 'Failed to search records',
+                details: error.message,
             });
-
-            const syncResults = [];
-
-            for (const person of attioPeople.data?.slice(0, args.maxPeople || 10) || []) {
-                try {
-                    // Transform person data for Quo
-                    const quoPersonData = await this.transformPersonForQuo(person);
-
-                    // Create or update in Quo if available
-                    let quoResult = null;
-                    if (this.quo?.api) {
-                        quoResult = await this.createQuoPerson(quoPersonData);
-                    }
-
-                    syncResults.push({
-                        attioPerson: {
-                            id: person.id,
-                            firstName: this.getPersonFirstName(person),
-                            lastName: this.getPersonLastName(person),
-                            email: this.getPersonEmail(person),
-                            companyName: this.getPersonCompany(person),
-                        },
-                        quoPerson: quoResult,
-                        syncStatus: quoResult ? 'success' : 'quo_unavailable',
-                        timestamp: new Date().toISOString(),
-                    });
-                } catch (personError) {
-                    syncResults.push({
-                        attioPerson: person,
-                        error: personError.message,
-                        syncStatus: 'error',
-                        timestamp: new Date().toISOString(),
-                    });
-                }
-            }
-
-            return {
-                label: 'People Sync Results',
-                data: {
-                    totalPeopleProcessed: syncResults.length,
-                    syncSummary: syncResults.reduce((summary, result) => {
-                        summary[result.syncStatus] = (summary[result.syncStatus] || 0) + 1;
-                        return summary;
-                    }, {}),
-                    syncResults,
-                    timestamp: new Date().toISOString(),
-                }
-            };
-        } catch (error) {
-            console.error('People sync failed:', error);
-            throw new Error(`People sync failed: ${error.message}`);
-        }
-    }
-
-    async createRecord(args) {
-        try {
-            const { objectType, attributes } = args;
-
-            if (!objectType) {
-                throw new Error('Object type is required');
-            }
-
-            let result;
-            switch (objectType.toLowerCase()) {
-                case 'company':
-                    result = await this.attio.api.createCompany(attributes);
-                    break;
-                case 'person':
-                    result = await this.attio.api.createPerson(attributes);
-                    break;
-                default:
-                    result = await this.attio.api.createRecord(objectType, attributes);
-            }
-
-            return {
-                label: `Created ${objectType} Record`,
-                data: {
-                    objectType,
-                    recordId: result.data?.id?.record_id,
-                    attributes: attributes,
-                    result,
-                    timestamp: new Date().toISOString(),
-                }
-            };
-        } catch (error) {
-            console.error('Failed to create record:', error);
-            throw new Error(`Failed to create record: ${error.message}`);
-        }
-    }
-
-    async searchRecords(args) {
-        try {
-            const { query, objectType, limit } = args;
-
-            let results = [];
-
-            if (objectType && objectType.toLowerCase() === 'company') {
-                const companies = await this.attio.api.searchCompanies(query, { limit: limit || 20 });
-                results = companies.data || [];
-            } else if (objectType && objectType.toLowerCase() === 'person') {
-                const people = await this.attio.api.searchPeople(query, { limit: limit || 20 });
-                results = people.data || [];
-            } else {
-                // Search across multiple object types
-                const [companies, people] = await Promise.all([
-                    this.attio.api.searchCompanies(query, { limit: limit ? Math.floor(limit / 2) : 10 }),
-                    this.attio.api.searchPeople(query, { limit: limit ? Math.floor(limit / 2) : 10 }),
-                ]);
-
-                results = [
-                    ...(companies.data || []).map(item => ({ ...item, _object_type: 'company' })),
-                    ...(people.data || []).map(item => ({ ...item, _object_type: 'person' })),
-                ];
-            }
-
-            return {
-                label: `Search Results for "${query}"`,
-                data: {
-                    query,
-                    objectType: objectType || 'all',
-                    totalResults: results.length,
-                    results: results.map(result => ({
-                        id: result.id,
-                        objectType: result._object_type || objectType,
-                        name: this.getRecordName(result),
-                        attributes: this.flattenRecordAttributes(result),
-                    })),
-                    timestamp: new Date().toISOString(),
-                }
-            };
-        } catch (error) {
-            console.error('Search failed:', error);
-            throw new Error(`Search failed: ${error.message}`);
-        }
-    }
-
-    // Helper methods to extract data from Attio's attribute structure
-    getCompanyName(company) {
-        return company.company_name?.values?.[0]?.value ||
-            company.company_name_alternative?.values?.[0]?.value ||
-            company.company_name_legal?.values?.[0]?.value ||
-            'Unknown Company';
-    }
-
-    getCompanyDomain(company) {
-        return company.company_domain?.values?.[0]?.value ||
-            company.company_website?.values?.[0]?.value ||
-            null;
-    }
-
-    getPersonFirstName(person) {
-        return person.person_first_name?.values?.[0]?.value ||
-            person.person_name_first?.values?.[0]?.value;
-    }
-
-    getPersonLastName(person) {
-        return person.person_last_name?.values?.[0]?.value ||
-            person.person_name_last?.values?.[0]?.value;
-    }
-
-    getPersonEmail(person) {
-        return person.person_email?.values?.[0]?.value ||
-            person.person_email_work?.values?.[0]?.value ||
-            person.person_email_personal?.values?.[0]?.value;
-    }
-
-    getPersonCompany(person) {
-        return person.company?.values?.[0]?.value ||
-            person.person_company?.values?.[0]?.value;
-    }
-
-    getRecordName(record) {
-        const objectType = record._object_type || 'unknown';
-
-        if (objectType === 'company') {
-            return this.getCompanyName(record);
-        } else if (objectType === 'person') {
-            const firstName = this.getPersonFirstName(record);
-            const lastName = this.getPersonLastName(record);
-            return `${firstName || ''} ${lastName || ''}`.trim() || 'Unknown Person';
-        }
-
-        return 'Unknown Record';
-    }
-
-    flattenRecordAttributes(record) {
-        const flattened = {};
-
-        for (const [key, value] of Object.entries(record)) {
-            if (key.startsWith('_') || !value) continue;
-
-            if (value.values && Array.isArray(value.values)) {
-                flattened[key] = value.values.map(v => v.value).join(', ');
-            } else if (typeof value === 'object') {
-                flattened[key] = JSON.stringify(value);
-            } else {
-                flattened[key] = value;
-            }
-        }
-
-        return flattened;
-    }
-
-    async transformCompanyForQuo(attioCompany) {
-        return {
-            name: this.getCompanyName(attioCompany),
-            domain: this.getCompanyDomain(attioCompany),
-            website: attioCompany.company_website?.values?.[0]?.value,
-            industry: attioCompany.company_industry?.values?.[0]?.value,
-            size: attioCompany.company_size?.values?.[0]?.value,
-            status: attioCompany.company_status?.values?.[0]?.value,
-            description: attioCompany.company_description?.values?.[0]?.value,
-            customFields: {
-                attioId: attioCompany.id,
-                source: 'Attio',
-                foundedYear: attioCompany.company_founded_year?.values?.[0]?.value,
-            },
-        };
-    }
-
-    async transformPersonForQuo(attioPerson) {
-        return {
-            firstName: this.getPersonFirstName(attioPerson),
-            lastName: this.getPersonLastName(attioPerson),
-            email: this.getPersonEmail(attioPerson),
-            title: attioPerson.person_title?.values?.[0]?.value,
-            phone: attioPerson.person_phone?.values?.[0]?.value,
-            company: this.getPersonCompany(attioPerson),
-            customFields: {
-                attioId: attioPerson.id,
-                source: 'Attio',
-                location: attioPerson.person_location?.values?.[0]?.value,
-            },
-        };
-    }
-
-    async createQuoCompany(companyData) {
-        if (!this.quo?.api) {
-            return { message: 'Quo API not available', status: 'simulated' };
-        }
-
-        try {
-            const quoCompany = await this.quo.api.createCompany(companyData);
-            return {
-                quoId: quoCompany.id || quoCompany.company_id,
-                name: companyData.name,
-                domain: companyData.domain,
-                status: 'created',
-            };
-        } catch (error) {
-            throw new Error(`Failed to create Quo company: ${error.message}`);
-        }
-    }
-
-    async createQuoPerson(personData) {
-        if (!this.quo?.api) {
-            return { message: 'Quo API not available', status: 'simulated' };
-        }
-
-        try {
-            const quoPerson = await this.quo.api.createPerson(personData);
-            return {
-                quoId: quoPerson.id || quoPerson.person_id,
-                name: `${personData.firstName} ${personData.lastName}`,
-                email: personData.email,
-                status: 'created',
-            };
-        } catch (error) {
-            throw new Error(`Failed to create Quo person: ${error.message}`);
-        }
-    }
-
-    async onCreate(params) {
-        this.record.status = 'ENABLED';
-        await this.record.save();
-        return this.record;
-    }
-
-    async onUpdate(params) {
-        await this.record.save();
-        return this.validateConfig();
-    }
-
-    async getConfigOptions() {
-        return {
-            syncInterval: {
-                type: 'number',
-                title: 'Sync Interval (minutes)',
-                description: 'How often to sync data between Attio and Quo',
-                default: 30,
-                minimum: 5,
-                maximum: 1440,
-            },
-            defaultObjectLimit: {
-                type: 'number',
-                title: 'Default Object Limit',
-                description: 'Default limit for listing objects',
-                default: 50,
-                minimum: 10,
-                maximum: 1000,
-            },
-        };
-    }
-
-    async getActionOptions({ actionId, data }) {
-        switch (actionId) {
-            case 'GET_ATTIO_CUSTOM_OBJECTS':
-                return {
-                    jsonSchema: {
-                        type: 'object',
-                        properties: {
-                            maxObjects: {
-                                type: 'number',
-                                title: 'Max Objects',
-                                description: 'Maximum number of custom objects to process',
-                                minimum: 1,
-                                maximum: 20,
-                                default: 5,
-                            },
-                            sampleLimit: {
-                                type: 'number',
-                                title: 'Sample Records Limit',
-                                description: 'Maximum number of sample records per object',
-                                minimum: 1,
-                                maximum: 10,
-                                default: 3,
-                            },
-                        },
-                        required: [],
-                    },
-                    uiSchema: {
-                        type: 'VerticalLayout',
-                        elements: [
-                            {
-                                type: 'Control',
-                                scope: '#/properties/maxObjects',
-                            },
-                            {
-                                type: 'Control',
-                                scope: '#/properties/sampleLimit',
-                            },
-                        ],
-                    },
-                };
-            case 'SYNC_COMPANIES_TO_QUO':
-                return {
-                    jsonSchema: {
-                        type: 'object',
-                        properties: {
-                            limit: {
-                                type: 'number',
-                                title: 'Company Limit',
-                                description: 'Maximum companies to retrieve for sync',
-                                minimum: 1,
-                                maximum: 1000,
-                                default: 50,
-                            },
-                            maxCompanies: {
-                                type: 'number',
-                                title: 'Max Companies to Sync',
-                                description: 'Maximum companies to actually sync',
-                                minimum: 1,
-                                maximum: 100,
-                                default: 10,
-                            },
-                            filter: {
-                                type: 'string',
-                                title: 'Filter (JSON)',
-                                description: 'Optional filter criteria in JSON format',
-                            },
-                        },
-                        required: [],
-                    },
-                    uiSchema: {
-                        type: 'VerticalLayout',
-                        elements: [
-                            {
-                                type: 'Control',
-                                scope: '#/properties/limit',
-                            },
-                            {
-                                type: 'Control',
-                                scope: '#/properties/maxCompanies',
-                            },
-                            {
-                                type: 'Control',
-                                scope: '#/properties/filter',
-                            },
-                        ],
-                    },
-                };
-            case 'SYNC_PEOPLE_TO_QUO':
-                return {
-                    jsonSchema: {
-                        type: 'object',
-                        properties: {
-                            limit: {
-                                type: 'number',
-                                title: 'Person Limit',
-                                description: 'Maximum people to retrieve for sync',
-                                minimum: 1,
-                                maximum: 1000,
-                                default: 50,
-                            },
-                            maxPeople: {
-                                type: 'number',
-                                title: 'Max People to Sync',
-                                description: 'Maximum people to actually sync',
-                                minimum: 1,
-                                maximum: 100,
-                                default: 10,
-                            },
-                            filter: {
-                                type: 'string',
-                                title: 'Filter (JSON)',
-                                description: 'Optional filter criteria in JSON format',
-                            },
-                        },
-                        required: [],
-                    },
-                    uiSchema: {
-                        type: 'VerticalLayout',
-                        elements: [
-                            {
-                                type: 'Control',
-                                scope: '#/properties/limit',
-                            },
-                            {
-                                type: 'Control',
-                                scope: '#/properties/maxPeople',
-                            },
-                            {
-                                type: 'Control',
-                                scope: '#/properties/filter',
-                            },
-                        ],
-                    },
-                };
-            case 'CREATE_ATTIO_RECORD':
-                return {
-                    jsonSchema: {
-                        type: 'object',
-                        properties: {
-                            objectType: {
-                                type: 'string',
-                                title: 'Object Type',
-                                description: 'Type of record to create',
-                                enum: ['company', 'person', 'opportunity', 'deal'],
-                                default: 'company',
-                            },
-                            attributes: {
-                                type: 'object',
-                                title: 'Attributes',
-                                description: 'Attributes for the record (JSON object)',
-                            },
-                        },
-                        required: ['objectType'],
-                    },
-                    uiSchema: {
-                        type: 'VerticalLayout',
-                        elements: [
-                            {
-                                type: 'Control',
-                                scope: '#/properties/objectType',
-                            },
-                            {
-                                type: 'Control',
-                                scope: '#/properties/attributes',
-                                options: {
-                                    format: 'textarea',
-                                    rows: 10,
-                                },
-                            },
-                        ],
-                    },
-                };
-            case 'SEARCH_ATTIO_RECORDS':
-                return {
-                    jsonSchema: {
-                        type: 'object',
-                        properties: {
-                            query: {
-                                type: 'string',
-                                title: 'Search Query',
-                                description: 'Search term or phrase',
-                                minLength: 1,
-                            },
-                            objectType: {
-                                type: 'string',
-                                title: 'Object Type',
-                                description: 'Specific object type to search (leave empty for all)',
-                                enum: ['', 'company', 'person', 'opportunity', 'deal'],
-                            },
-                            limit: {
-                                type: 'number',
-                                title: 'Result Limit',
-                                description: 'Maximum number of results to return',
-                                minimum: 1,
-                                maximum: 100,
-                                default: 20,
-                            },
-                        },
-                        required: ['query'],
-                    },
-                    uiSchema: {
-                        type: 'VerticalLayout',
-                        elements: [
-                            {
-                                type: 'Control',
-                                scope: '#/properties/query',
-                            },
-                            {
-                                type: 'Control',
-                                scope: '#/properties/objectType',
-                            },
-                            {
-                                type: 'Control',
-                                scope: '#/properties/limit',
-                            },
-                        ],
-                    },
-                };
-            default:
-                return null;
         }
     }
 }
